// src/app/interfaces/persona/persona-modal.ts
export interface PersonaModal {
    idPersona?: string,
    primerNombreModal?: string;
    segundoNombreModal?: string;
    apellidoPaternoModal?: string;
    apellidoMaternoModal?: string;
    idTipoDocumentoModal?: string;
    tipoDocumentoModal?: string;
    numeroDocumentoModal?: string;
    idTipoTelefonoModal1?: string;
    idTelefonoModal1?: string;
    telefonoModal1?: string;
    idTipoTelefonoModal2?: string;
    idTelefonoModal2?: string;
    telefonoModal2?: string;
    idCorreoModal?: string;
    correoModal?: string;
    idComunidadModal?: string;
    comunidadModal?: string;
    idParroquiaModal?: string;
    parroquiaModal?: string;
    idCantonModal?: string;
    cantonModal?: string;
    idProvinciaModal?: string;
    provinciaModal?: string;
    idAsignacionPC?: string;
}